Tidy up source creation for suite tests
